+++ src/hex/roller.ts
@@ -31,7 +31,7 @@ export class HEXRoller extends Roller<Dice, Faces, DicePool> {
         super(command, [new SimpleParser()], true, true);
     }
 
-    public roll(pool: DicePool): Array<Roll<Dice, Faces>> {
+    public roll(pool: DicePool): Roll<Dice, Faces>[] {
         if (pool.BONUS >= pool.MALUS) {
           pool.BONUS = pool.BONUS - pool.MALUS;
           pool.MALUS = 0;
@@ -50,7 +50,7 @@ export class HEXRoller extends Roller<Dice, Faces, DicePool> {
         ];
     }
 
-    public combineRolls(rolls: Array<Roll<Dice, Faces>>): RollValues {
+    public combineRolls(rolls: Roll<Dice, Faces>[]): RollValues {
         const results = rolls
             .map((roll) => parseRollValues(roll));
         return combineAll(results, rollValuesMonoid);
@@ -60,7 +60,7 @@ export class HEXRoller extends Roller<Dice, Faces, DicePool> {
         return new Roll(die, face);
     }
 
-    public formatRolls(rolls: Array<Roll<Dice, Faces>>, flavorText?: string): string {
+    public formatRolls(rolls: Roll<Dice, Faces>[], flavorText?: string): string {
         const combinedRolls = combineRolls(rolls, parseRollValues, rollValuesMonoid);
         return Mustache.render(
             base,