+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.12.0] - 2020-10-02
+### Added
+- WFRP3 roller, #10, thanks @silentmark, @illotum!
+
 ## [0.11.3] - 2020-06-13
 ### Fixed
 - Do not log undefined to chat console if no roller command was found